+++ flash_cards/flash_cards/View/EditWordView.swift
@@ -8,19 +8,6 @@
 import SwiftUI
 import ComposableArchitecture
 
-// 単語を表す構造体
-struct Word: Identifiable, Equatable {
-    var id: UUID
-    var term: String
-    var definition: String
-    
-    init(id: UUID = UUID(), term: String, definition: String) {
-        self.id = id
-        self.term = term
-        self.definition = definition
-    }
-}
-
 // 単語編集画面の状態
 struct EditWord: Reducer {
     struct State: Equatable {
